Test registro productos en base de datos
